Create UNITE_HAB_2_0
This was my attempt to interface the base HAB_UNITE code with the magnetometer code without having an Arduino or magnetometer to work with. The only problem I anticipate is that the mag variables are set to be bytes in the code (which they will need to be for transmission) but they are possibly two bytes based on magnetometer reading code. We will need to fix this. 
-Adam